+++ examples/README_example3.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from plotez import n_plotter
-from plotez.backend import LinePlotConfig, SubPlotConfig
+from plotez.backend import FigureConfig, LinePlotConfig
 
 # Create 2×2 grid
 x_data = [np.linspace(0, 10, 100) for _ in range(6)]
@@ -18,10 +18,10 @@ y_data = [
 ]
 
 line_config = LinePlotConfig(color=["red", "blue", "green", "black", "orange", "magenta"])
-subplot_config = SubPlotConfig(figsize=(10, 6))
+figure_config = FigureConfig(figsize=(10, 6))
 
 fig, axs = n_plotter(
-    x_data, y_data, n_rows=2, n_cols=3, auto_label=True, plot_config=line_config, subplot_config=subplot_config
+    x_data, y_data, n_rows=2, n_cols=3, auto_label=True, plot_config=line_config, figure_config=figure_config
 )
 
 # plt.show()