+++ src/plotez/backend/CONSTANTS.py
@@ -9,11 +9,16 @@ LINE_ATTRS = {
     "mfc": "markerfacecolor",
     "mew": "markeredgewidth",
 }
+
 ERROR_ATTRS = LINE_ATTRS | {
     "ecolor": "ecolor",
     "elinewidth": "elinewidth",
     "capsize": "capsize",
     "capthick": "capthick",
 }
+
 SCATTER_ATTRS = {"c": "color", "s": "size", "ec": "edgecolors", "fc": "facecolors"}
+
 SUBPLOT_ATTRS = {"sharex": "sharex", "sharey": "sharey", "figsize": "figsize"}
+
+ERROR_BAND_ATTRS = {"color": "color"}