+++ src/plotez/backend/__init__.py
@@ -1,12 +1,13 @@
 """Created on Jul 20 00:17:08 2022."""
 
-from .CONSTANTS import ERROR_ATTRS, LINE_ATTRS, SCATTER_ATTRS, SUBPLOT_ATTRS
+from .CONSTANTS import ERROR_ATTRS, ERROR_BAND_ATTRS, LINE_ATTRS, SCATTER_ATTRS, SUBPLOT_ATTRS
 from .error_handling import OrientationError
 from .utilities import (
+    ErrorBandConfig,
     ErrorPlotConfig,
+    FigureConfig,
     LinePlotConfig,
     ScatterPlotConfig,
-    SubPlotConfig,
     dual_axes_data_validation,
     dual_axes_label_management,
     plot_or_scatter,
@@ -20,9 +21,11 @@ __all__ = [
     "SUBPLOT_ATTRS",
     "OrientationError",
     "ErrorPlotConfig",
+    "ErrorBandConfig",
+    "ERROR_BAND_ATTRS",
     "LinePlotConfig",
     "ScatterPlotConfig",
-    "SubPlotConfig",
+    "FigureConfig",
     "dual_axes_data_validation",
     "dual_axes_label_management",
     "plot_or_scatter",